backend/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
openai
pyautogen
python-dotenv